+++ packages/DataField/CHANGELOG.md
@@ -1,5 +1,12 @@
 <!-- content -->
 
+## 1.0.26-next.0
+
+### Patch Changes
+
+- Updated dependencies [371d3d3]
+  - @paprika/popover@2.0.0-next.0
+
 ## 1.0.25
 
 ### Patch Changes