+++ packages/OverflowMenu/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 2.0.0-next.0
+
+### Major Changes
+
+- 371d3d3: Refactored Popover to not add content to the dom when closed by default
+
+### Patch Changes
+
+- Updated dependencies [371d3d3]
+  - @paprika/popover@2.0.0-next.0
+  - @paprika/confirmation@1.1.14-next.0
+
 ## 1.1.9
 
 ### Patch Changes